fix(docs): unblock luaJavadoc on new Create peripherals
cct-javadoc rejects float returns, {@value} tags, and {@link} to
LuaException. Widen four float Lua getters to double, replace two
{@value #MAX_INSTRUCTIONS} with the literal, and demote one
{@link LuaException} to plain text.

+++ common/src/main/java/ink/astrius/create_avionics/compat/create/peripherals/SpeedControllerPeripheral.java
@@ -72,7 +72,7 @@ public class SpeedControllerPeripheral extends KineticPeripheral<SpeedController
     }
 
     @LuaFunction
-    public final float getTargetSpeed() {
+    public final double getTargetSpeed() {
         return this.targetSpeed.getValue();
     }
 }
